Implemented Sign Up and Sign In
A number of files were created to handle the api calls and event
handling.  Also the initial html was created to handle these features.

This code was implemented using methods discussed in lectures.

It was determined that the html for each feature(i.e. Sign In) required
that the html be wrapped in a form. The id associated with this form
is used in the event handler to get that node(form) so the node elements
can be parsed and formatted as expected by the api calls.

Had an issue with the bootstrap button where it would not work using
'submit'. Switch to method input and worked as expected.