When Sign Out disable board and clean up
Also updated jquery call when using an ID need to build the string
inside $() when enabling events and disabling.

If an event is called more than expected ensure there is not
another being added that is not expected.  There can be more than
one event added to an element.   s